Created a utility function for scrollKeyboardAwaring View
