+++ src/app/(auth)/forget-password/forget-password-otp.tsx
@@ -2,7 +2,7 @@ import { ErrorDialog } from "@/components/error-dialog";
 import { P } from "@/components/ui/typography";
 import { useLocalSearchParams } from "expo-router";
 import { useEffect, useState } from "react";
-import { KeyboardAvoidingView, Platform, ScrollView } from "react-native";
+import { KeyboardAwareScrollView } from "@/components/keyboard-aware-scrollView";
 
 import { ForgotPasswordOTPBanner } from "@/modules/auth/forget-password/components/forgotPassword/forgot-password-banner";
 // Forget Password Otp Input Screen Components
@@ -28,20 +28,7 @@ export default function ForgetPasswordOTPScreen() {
 		if (error) setDialogOpen(true);
 	}, [error]);
 	return (
-		<KeyboardAvoidingView
-			behavior={Platform.OS === "ios" ? "padding" : "height"}
-			keyboardVerticalOffset={Platform.OS === "ios" ? 100 : 0}
-			style={{ flex: 1 }}
-		>
-			<ScrollView
-				contentContainerStyle={{
-					flexGrow: 1,
-					padding: 8,
-					justifyContent: "center",
-					alignItems: "center",
-					rowGap: 32,
-				}}
-			>
+		<KeyboardAwareScrollView className="justify-center items-center p-2 gap-y-8">
 				{/* Forgot Password OTP Banner */}
 				<ForgotPasswordOTPBanner
 					phoneNumber={phoneNumber}
@@ -66,7 +53,6 @@ export default function ForgetPasswordOTPScreen() {
 					open={dialogOpen}
 					onOpenChange={setDialogOpen}
 				/>
-			</ScrollView>
-		</KeyboardAvoidingView>
+		</KeyboardAwareScrollView>
 	);
 }
