+++ src/app/(auth)/login/index.tsx
@@ -1,14 +1,9 @@
 import { ErrorDialog } from "@/components/error-dialog";
+import { KeyboardAwareScrollView } from "@/components/keyboard-aware-scrollView";
 import { H1, Muted } from "@/components/ui/typography";
 import { Link } from "expo-router";
 import React, { useEffect, useState } from "react";
-import {
-	KeyboardAvoidingView,
-	Platform,
-	ScrollView,
-	useWindowDimensions,
-} from "react-native";
-import { useSafeAreaInsets } from "react-native-safe-area-context";
+import { useWindowDimensions } from "react-native";
 
 // Login Screen Components
 import { LoginBanner } from "@/modules/auth/login/components/loginForm/loginBanner";
@@ -22,8 +17,6 @@ export default function LoginScreen() {
 	const { mutate, error, isPending } = useloginUserForm();
 	const [dialogOpen, setDialogOpen] = useState(false);
 
-	const safeAreaInsets = useSafeAreaInsets();
-
 	const windowDimensions = useWindowDimensions();
 
 	useEffect(() => {
@@ -31,57 +24,39 @@ export default function LoginScreen() {
 	}, [error]);
 
 	return (
-		<KeyboardAvoidingView
-			behavior={Platform.OS === "ios" ? "padding" : "height"}
-			keyboardVerticalOffset={Platform.OS === "ios" ? 100 : 0}
-			style={{
-				flex: 1,
-				paddingBottom: safeAreaInsets.bottom,
-				paddingTop: safeAreaInsets.top,
-			}}
-		>
-			<ScrollView
-				contentContainerStyle={{
-					flexGrow: 1,
-					padding: 8,
-					justifyContent: "center",
-					alignItems: "center",
-					rowGap: 20,
-				}}
+		<KeyboardAwareScrollView className="justify-center items-center gap-y-5">
+			{/* Login Banner */}
+			{windowDimensions.height > 700 ? <LoginBanner /> : undefined}
+
+			<H1>Login</H1>
+
+			{/* Login Form */}
+			<LoginForm isLoading={isPending} triggerLogin={mutate} />
+
+			{/* Error Component */}
+			{error && (
+				<ErrorDialog
+					error={error}
+					open={dialogOpen}
+					onOpenChange={setDialogOpen}
+				/>
+			)}
+			{/* Link to forget Password */}
+			<Link
+				href={"/(auth)/forget-password"}
+				className="text-black dark:text-white"
 			>
-				{/* Login Banner */}
-				{windowDimensions.height > 700 ? <LoginBanner /> : undefined}
+				Forgot Password ?
+			</Link>
 
-				<H1>Login</H1>
+			<Muted className="text-lg">Or Continue With</Muted>
 
-				{/* Login Form */}
-				<LoginForm isLoading={isPending} triggerLogin={mutate} />
+			{/* Social Provider Login */}
+			<SocialProviderLogin />
 
-				{/* Error Component */}
-				{error && (
-					<ErrorDialog
-						error={error}
-						open={dialogOpen}
-						onOpenChange={setDialogOpen}
-					/>
-				)}
-				{/* Link to forget Password */}
-				<Link
-					href={"/(auth)/forget-password"}
-					className="text-black dark:text-white"
-				>
-					Forgot Password ?
-				</Link>
-
-				<Muted className="text-lg">Or Continue With</Muted>
-
-				{/* Social Provider Login */}
-				<SocialProviderLogin />
-
-				<Link href={"/(auth)/register"} className="text-black dark:text-white">
-					Don't have an account
-				</Link>
-			</ScrollView>
-		</KeyboardAvoidingView>
+			<Link href={"/(auth)/register"} className="text-black dark:text-white">
+				Don't have an account
+			</Link>
+		</KeyboardAwareScrollView>
 	);
 }
